程嘉朗

+++ src/main/java/com/example/projecthelper/service/AssignmentService.java
@@ -64,8 +64,10 @@ public class AssignmentService {
             }
             results = assignmentMapper.getAssByProj(projId, pageSize, page * pageSize);
         }
-        results.forEach(a ->
-            a.setFilePaths(a.getFilePaths().stream().map(FileUtil::getFilenameFromPath).toList())
+        results.forEach(a ->{
+                if(a.getFilePaths() != null)
+                    a.setFilePaths(a.getFilePaths().stream().map(FileUtil::getFilenameFromPath).toList());
+            }
         );
         return results;
     }
@@ -252,10 +254,14 @@ public class AssignmentService {
             submittedAssMapper.removeAss(assignmentId, userId);
         }
         else if (assignment.getType().equals("g")){
+
             Long gpId = groupMapper.findGroupIdOfUserInAProj(userId, assignment.getProjectId());
             if(gpId != null){
-                fileService.removeFilesOfSubmittedAss(assignment, gpId);
-                submittedAssMapper.removeAss(assignmentId, gpId);
+                if(!Objects.equals(userId, groupMapper.findLeaderByGroup(gpId))){
+                    fileService.removeFilesOfSubmittedAss(assignment, gpId);
+                    submittedAssMapper.removeAss(assignmentId, gpId);
+                    return;
+                }
             }
             throw new AccessDeniedException("你不在小组中");
         }
